update constants.js to TS update register-slice.js to register-slice.ts
